# machi-openapi/openapi.v1.yaml
openapi: 3.0.0
info:
    title: User
    version: "1.0"
servers:
    - url: http://localhost:3000
paths:
    "/user":
        post:
            summary: ""
            operationId: post-user
            responses:
                "201":
                    description: OK
                    content:
                        application/json:
                            schema:
                                "$ref": "#/components/schemas/ResponseUser"
                "400":
                    description: Bad Request
                    content:
                        application/jaon:
                            schema:
                                "$ref": "#/components/schemas/ResponseUser"
                "404":
                    description: Not Found
            requestBody:
                content:
                    application/json:
                        schema:
                            "$ref": "#/components/schemas/UserDraft"

    "/user/{id}":
        parameters:
            - schema:
                  type: number
              name: id
              in: path
              required: true
        put:
            summary: update user
            tags: []
            responses:
                "200":
                    description: OK
                    content:
                        application/jaon:
                            schema:
                                "$ref": "#/components/schemas/ResponseUser"
                "400":
                    description: Bad Request
                    content:
                        application/jaon:
                            schema:
                                "$ref": "#/components/schemas/ResponseUser"
                "404":
                    description: Not Found
            operationId: put-user-id
            requestBody:
                content:
                    application/json:
                        schema:
                            "$ref": "#/components/schemas/User"
        delete:
            summary: remove user
            tags: []
            responses:
                "204":
                    description: No Content
                "404":
                    description: Not Found
            operationId: delete-user-id
        get:
            summary: get user
            tags: []
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                "$ref": "#/components/schemas/User"
                "404":
                    description: Not Found
            operationId: get-user-id
    "/users":
        get:
            summary: Your GET endpoint
            tags: []
            responses:
                "200":
                    description: OK
                    headers: {}
                    content:
                        application/json:
                            schema:
                                "$ref": "#/components/schemas/User"
                            examples: {}
            operationId: get-user
        parameters: []
    "/users/find":
        post:
            summary: ""
            operationId: post-users-find
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                "$ref": "#/components/schemas/User"
            requestBody:
                content:
                    application/json:
                        schema:
                            "$ref": "#/components/schemas/UserQuery"
components:
    schemas:
        ResponseError:
            "$ref": "./models/ResponseError.yaml#/"
        ResponseUser:
            "$ref": "./models/ResponseUser.yaml#/"
        User:
            "$ref": "./models/User.v1.yaml#/"
        Users:
            "$ref": "./models/Users.v1.yaml#/"
        UserDraft:
            "$ref": "./models/UserDraft.v1.yaml#/"
        UserQuery:
            "$ref": "./models/UserQuery.v1.yaml"